DB‑MS/requirements.txt
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
Flask==2.2.3
flask-cors==5.0.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==3.2.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.3.0
PyJWT==2.6.0
python-dotenv==0.21.1
requests==2.28.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==1.26.20
Werkzeug==2.2.3